[workflow] Bugfix.
Found a bug in experiment.py in which the same pulse was used for all the simulations. This was due to lambda expressions and passing objects by reference. Fixed the bug by passing the object to the lambda as default value, such that it is evaluated on instantiation of the lambda function.